[PE] Add ToInt() for logical padding top and bottom for LayoutTableCell
This reverts the logic for logical top/bottom paddings to be the same
as before https://chromium-review.googlesource.com/619603. The logical
top/bottom paddings are truncated like what we do for other table
layouts.

We still keep subpixels for logical left/right paddings to ensure
correct logical width calculation. This is still the same as before
https://chromium-review.googlesource.com/619603.

Bug: 823824
Change-Id: Icb617bf5771d1726f767207853f4adc8275731dc